requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
easyocr>=1.7.1
numpy>=1.24.0
pandas>=2.0.0
twilio>=8.10.0
python-dotenv>=1.0.0
Pillow>=10.0.0
